style(black): lint.ai auto-Blackifier
Lint.ai periodically searches for code that could benefit from automatic
formatting and applies best-in-breed code reformatting courtesy of the
Black project (github.com/ambv/black)*.  This serves both to standardize
the formatting of this project's code, and to test Black itself; we
enable a Black mode that guarantees the compiled version of the
reformatted code *exactly* equals the compiled version of the original
code, so you can be confident that this reformatting introduces no bugs.

If you find this useful, please consider merging this pull request.  If
you do not, please close it; as part of our Helpful Robot Policy, we
will never open another PR if our first was not merged.

*: Please note that Lint.AI is *not* controlled by the author of Black.

ABUSE: Send complaints/suggestions/hate mail/photographs of caterpillars
       to blackifier@lint.ai.