Add '-' param to read from stdin
Tools like ALE and so forth generally pass the buffer of the editor
directly to the tool to edit and expect to read it back on stdout. I've
added the `-` to signal to add-trailing-comma to read data from stdin
and output to stdout.

I modeled this mostly off of black. Black doesn't have any error
handling around passing multiple files and `-`. Black also only briefly
mentions it in the docs.

I've also added a basic tests to check that a (mocked) stdin works.